Add nova_api_metadata_local_check.py script
This is a basic script to test the availability of a nova-api-metadata
service.  Unfortunately we can only get rather limited data back (in
this case, ec2 metadata versions) without specifying a bunch of headers
(such as existing IP address, etc.)

Note that I couldn't seem to get json back, so I have to parse the
string content looking for a particular version.