Monitor swift with monitoring as a service
Most of the desired monitoring output is generated by swift-recon. To
simplify the structure of swift monitors, they're all in one script and
the appropriate output will be printed based on the command desired.
This can also be used as a module since the functions that parse the
swift-recon output is separate from what prints it.

Example usage::

    python swift-recon.py replication --ring object
    python swift-recon.py async-pendings

This will check replication, md5sums, and quarantined objects without
extra configuration. In order to collect information about the pending
async operations, a cronjob needs to be created on the object nodes.
See: http://docs.openstack.org/developer/swift/admin_guide.html#cluster-telemetry-and-monitoring