Removed hard set requirements
This change removes the hard set requirements which cause conflicts
with the ansible-lxc-rpc repo. Because the packages are locked
down independently in the ansible-lxc-rpc these package entries
have a high probability of creating a conflict. Any version
of the monitoring repo can be deployed with any version the
ansible-lxc-rpc which results in conflict as the newest version of
the package will always be installed.