Import rpc_maas_tool.py from os-ansible-deployment
This commit creates a scripts directory similar to
os-ansible-deployment and imports rpc_maas_tool.py from master in
os-ansible-deployment prior to it getting removed in the Rackspace
cleanup.  We also change the script name to rpc-maas-tool.py for
consistency and add the additional 'success' status in the check
method.  The latter change will be backported to juno/icehouse.

Fixes #25