Dynamically determine hostname for MaaS checks
Currently, we assume hosts returned from a `nova service-list` and
`neutron agent-list` are FQDNs which is an incorrect assumption. The
original change was made as we ran these MaaS-related playbooks/roles
on hosts w/ incorrectly configured hostnames (meaning a `hostname`
would return the FQDN), and therefore the hosts being returned from the
above two commands were FQDNs.

This change replaces references to ansible_fqdn to ansible_nodename,
which uses a python platform.node() to obtain the hostname. This should
allow us to properly configure MaaS irrespective of whether or not an
installation has correctly configured hostnames on the individual nodes.

Closes issue #9