Fix regression introduced in adcffa0
In commit adcffa0, we added |default('') to maas_fqdn_extension since
maas_fqdn_extension was not defaulted in any user variable files and
would cause an ansible run failure due to the variable not existing.
Upon doing this, I noticed that maas_fqdn_extension was _always_ being
set to '', even when it was defined in a user variables file.  Oddly,
I could print the contents of maas_fqdn_extension right above the
change made and it was visible, so it is unclear why adding the
|defauilt('') was causing this issue.

This change removes any |default('') reference on maas_fqdn_extension
and sets a default value in rpc_maas/defaults/main.yml instead. This is
the better solution and should have been done in the first place.

Closes #30