Add hp_monitoring.py script
The hp_monitoring.py script adds basic hardware monitoring for HP boxes
and is designed to complement the openmanage.py script for Dell
hardware.

The script makes use of the utilities hpasmcli and hpssacli. Details of
the repo from which to get them can be found at
https://downloads.linux.hp.com/SDR/project/mcp/. The packages installed
were hp-health and hpssacli.