Add F5 pool monitoring scripts
The scripts f5-config.py and f5-monitor.sh are part of the juno branch
of os-ansible-deployment. The were removed in master because they are
Rackspace specific files.

This commit adds these scripts here so they are available for the kilo
release.

f5-monitor.sh was modified from the juno version, this was done because
keystone no longer supports XML. The script now parses the JSON
response for the auth token.

f5-config.py was modified from the juno version; references to
rpc_deploy and rpc_inventory have been replaced with openstack_deploy
and openstack_inventory.

Closes-Bug: https://github.com/rcbops/rpc-extras/issues/46