Use nova API v2 instead of v2.1
In commit 347463d we changed the nova plugins to use v2.1 instead of
v3.  This was done as v3 is now deprecated and defaults to being
disabled in os-ansible-deployment.  The problem was the testing was
done on a deployment w/ v3 enabled and it appears there is some
coupling between v2.1 and v3.

This commit simply updates the nova plugins to use v2 instead of v2.1.
Thanks to @miguelgrinberg for spotting this!

Closes issue #197