Add distributed, multi-line log shipping
This patch adds the beaver role and playbook to rpc-extras. Beaver is a log
shipper which allows log shipping to be configured to squash multi-line log
entries into one event prior to shipping them to logstash. It also enables
the addition of tags and fields at the source. Each of these capabilities
distributes some of the processing currently handled by logstash to the
clients, improving the ability for RPC to handle logging at a greater scale.

Beaver is now set to be the log shipper used by default for RPC and is
executed within the setup-logging playbook.

This patch revises the logstash role to be a lot simpler, primarily to improve
log processing performance but also to be more standard with similar efforts
elsewhere in the OpenStack community.

To ensure that the Kibana dashboard reflects the changes in the logstash
filters, it is also adjusted. The default event window has also been changed
from 5m to 15m.

Changes affecting upgrades from previous versions:
 - Logstash now listens for normal syslog messages on port 5544 instead of
   json formatted syslog messages.
 - The user variable logstash_port is now logstash_syslog_port in order to
   differentiate it from the tcp port which beaver uses.
 - For easier troubleshooting any messages coming into logstash via syslog
   or tcp are tagged as logstash-input-syslog and logstash-input-tcp
   respectively.
 - Logstash field changes to standardise with other OpenStack Logstash
   efforts made in the community:
    - httpversion > http_ver
    - @message > message
    - openstack_code > (removed)
    - openstack_function > module
    - os_level > loglevel
    - os_program_path > (removed)
    - os_program_pid > pid
    - os_timestamp > @timestamp / req_data & req_time
    - request > url_path
    - request_id > requestid
    - request_user_id > user_id
    - request_tenant_id > tenant
 - New log files processed:
    - auth
    - beaver
    - elasticsearch
    - logstash
    - mysql
    - rabbitmq
 - Multline processing enabled for:
    - elasticsearch
    - logstash
    - mysql-slow log
    - OpenStack TRACE log lines
    - rabbitmq