Update rabbitmq_status.py
This commit updates rabbitmq_status.py to check /api/connections
instead of /api/channels.  According to rabbitmqctl, 'numbers'
on /api/channels is:

The number of the channel, which uniquely identifies it within a
connection.

Alarming if number > 1 is ineffective, and on OpenStack Kilo this
actually poses a problem as doing a 'rabbitmqctl list_channels
numbers' actually returns channels w/ name > 1.

Further, in https://bugs.launchpad.net/oslo.messaging/+bug/1406629 we
see the following command being used to determine problem state:

rabbitmqctl list_connections name channels | awk '{ if ($4 >1) print }'

This confirms that we should be checking /api/connections and not
/api/channels.

Closes issues #188