Repo build for rpc-extras
This patch adds the capability for rpc-extras to build wheels and
store them on the local repo server. The repo-build play scans
through all of the YAML files in the tree and creates wheels for
any packages or git sources it finds and places the wheels in the
rpc-extras tree of the repo server. The repo-pip-setup play adds
the rpc-extras tree to the pip configuration of all hosts. Due to
the way repo-build scans empty YAML files are not allowed, hence
the removal of the haproxy main.yml file.

Updated rpc_release comment and changed dependency scope.

(cherry picked from commit d0a8374792db6d5f0a514e53f772fba7a7ef0da2)

	README.md