Revert "create entities if missing"
This change reverts 096b0a2 as we originally made a conscious decision
to not create entities via rpc_maas. These entities should be created at
provisioning time and a lack of existence could indicate a problem with
host provisioning. Also, the rpc_maas role fails to run with this commit
in place. It looks like the entity_count var gets re-registered even
when the task is skipped resulting in there being no stdout field. In
this case it would probably be best to just re-issue the count task
again without a when condition.

This reverts commit 096b0a2a2bf679e326bbc9b75933c901ca3241c5.