Add user_variable.yml file to rpcd
This repo is for deploying an opinionated version of OpenStack using
os-ansible-deployment. There are instances where the defaults that a
Rackspace deployment will use differ from the upstream deployment
project. This commit creates a user_variables.yml file to use for
specifying these defaults and so this file should be used instead of
the one in os-ansible-deployment.

The user_variables.yml file in os-ansible-deployment currently sets two
variables that are not specified by default, by the associated roles.
These variables, ssl_protocol and ssl_cipher_suite, are specified in
this new file.

Support have requested that the nova configuration option
[cinder]/cross_az_attach be set to False and so is set here to override
the upstream default of True.

The script update-yaml.py is added, for use when running an AIO, so
that the user_variables file added by this commit can be merged into
the one copied by bootstrap-aio.sh. Variables defined in the rpcd
user_variables.yml file will override those in the one supplied by
bootstrap-aio.sh.

Closes-bug: https://github.com/rcbops/rpc-openstack/issues/203