add script to check if a given commit is contained in a built wheel
It's somewhat difficult to determine if a deployed wheel of an
openstack project (say neutron) contains a particular upstream commit,
because of the way we build wheels outside of the containers.

This commit adds a script to be run from the deployment host that does
the following things:

1) connects to all containers of type 'project' and determines the sha
from which the wheel was built
2) clones git repo locally (/tmp by default) to allow interrogation of
git history
3) checks to see if 'supplied commit sha' (the commit that you are
interested in) is contained in the wheel.

optional:
- supply a parent sha and use that instead of getting the
parent sha from the wheel in the containers.
- supply a host group instead of defaulting to <PROJECT>_all for the
  ansible run

Closes: #195
(cherry picked from commit 0c728a71afeadc9bc636023c04301fdb09933547)