Add MaaS specific rabbitmq user
Since we are removing the guest/guest user from rabbitmq as part of
os-ad we need to create a maas specific user that has "administrator"
privileges, otherwise the monitors don't work.

This patch adds a maas_user to rabbit, with a password specified in
user_extras_secrets. This is then passed to the monitors for rabbit as
they are created.

This won't adjust existing MaaS checks, so existing checks will need to
be removed and re-added or manually adjusted to address this problem.

Fixes: #223
(cherry picked from commit 6fc27ad7fd5a0b0191adc366faa91f8261f3c50e)