Add FORKS env variable to deploy
Currently the FORKS variable is set by osads scripts-library
to 25. This can cause issues with multinode deployments and the
ssh service in containers not being up in time for plays.

This will allow the user of the deploy script to override the
FORKS ansible variable to suite their environment. Defaults to the
number of CPUs of the deployment machine.