set uri correctly for maas checks by default
It seems that by default we are enabling ssl for horizon (self signed)
because of this we need to set the default uri for checks to be https
or maas will get confused (the check already defaults to port 443 (hard
coded)).

we also set the path to work around maas failures on redirects (from
master).

fixes #249