add 'version' param to get_url_endpoint_for_service method
Sometimes we have multiple versions of a particular endpoint type for a service.  This change allows us to pass a 'version' which is the last part of the url. e.g. for keystone we have both the v2 and v3 endpoints deployed, and in the catalog they look like this:

http://172.29.236.3:35357/v3
http://172.29.236.3:35357/v2.0

To ensure we return the v3 endpoint, you'd pass version='v3', to get the v2 endpoint, you'd pass version='v2.0'

The default bahaviour is retained, which is to pass no version and return the first match

Closes: #322
(cherry picked from commit 538d33fc872548abd465824189bc50bf3508510b)