Add ELK support for Ceph
Add logstash template file for ceph.
Add a beaver ceph.conf file for ceph containers and hosts.

Due to the permissions on /var/log/ceph.log and /var/log/ceph-audit.log
it isn't possible to capture these at this point, and Ceph offers no
otpions to adjust this, so we will need an alternative solution for
this. This applies to the mons only.