Add plugin and check/alarm to monitor volume group usage
Plugin:
Takes vg name as an argument and reports the following metrics:

vg_total_space (Mb)
vg_free_space (Mb)
vg_used_space (Mb)

Also included is a config file to set up the check, and
warning/critical alarms on (default) 80/90% used.

The playbooks are run in such a way that the check should only be added
on a cinder_volume host where the cinder backend is lvm.