Re-template raxrc file hosts for MaaS upgrade
When users first deploy MaaS they have to place a token in the correct
variables file. Tokens expire quickly so when they go to upgrade the
MaaS checks, then they'll need to upgrade their MaaS token again. We
need to ensure the token in the variable file is then placed on the
hosts in /root/.raxrc. This allows rpc-maas-tool.py to work as expected
within the playbooks.

Addresses #543