Convert pip tasks to support pip_install_options
In testing an upgrade from 10.1.15 to r11.0.x, @toxick found that some
packages installed on the host for MaaS need to be forcibly reinstalled
to match requirement constraints on newer packages. The case that
spurred this fix and the original issue was oslo.utils. The exact
information is documented in an issue comment.

Addresses #549

Conflicts:
	rpcd/playbooks/roles/rpc_support/tasks/pccommon_get.yml

(cherry picked from commit e7c786fc93ff0b69521bd3fa614633b54731ed5b)