Use 169.254.169.254 for neutron metadata check
The plugin uses the IP address 127.0.0.1 for monitoring the metadata
service availability, this is not the IP address used by nova instances
when accessing metadata and there have been reports of plugin failures
with this address.

This commit changes the IP address to 169.254.169.254 which should
increase the reliability of this plugin.

Closes-issue: #569