Search for the current host in nodes
To ensure we find the current host we should search for the hostname
without domain in the response array of nodes, and then proceed to
collect metrics from that dict.

We ensure we check that the RabbitMQ version is at least 3.5.x before
checking the lengths of queues on all nodes since we know that this
check is not valid in version 3.3.5 and further noodling into the
version where that feature becomes available can be deferred based on
need.

Addresses: #610