Update fail messages for MaaS entity discovery
Entities should alway exist for the customers hardware and are created
when it is assigned to the account. This commit updates the error
messages generated to be clearer about the action support should take.

The use of entites is not RPC specific and so no additional
documentation should be required.

Closes-issue: https://github.com/rcbops/rpc-openstack/issues/155