Bump ceph-common and ceph-mon SHAs
This commit bumps the ceph-common and ceph-mon role SHAs to their
latest commits. This fixes an issue where a fetch of a file from a
remote host to local host was running across all MON nodes
simultaneously, resulting in a potential race condition.

Note that the fix above is only in ceph-mon, however additional
changes to ceph-mon require that ceph-common gets bumped also.

Closes issue #599