Add Ansible linting under Ansible 2.x
This commit adds a new tox environment for linting playbooks and roles
against Ansible 2.x. It also modifies the linting-ansible.sh script so
that it is compatible with both the 1.x and 2.x series.

To support easy removal of the job in the future, the version of ansible used
is implemented as an environment variable that is substituted by tox. Thus,
making ansible 2.x the default in the future is as simple as removing
the line in .travis.yml that sets the environment variable to >=2.0 and
updating the default value on the requirement in tox.ini.

Also, the requirements files were changed from referencing the
openstack-ansible submodule to using the test-requirements.txt file
in-tree.

This commit is not intended to provide full Ansible 2.x support. Rather,
it is to serve as an early warning system so that the playbooks and
roles aren't written in an incompatible manner prior to actually
switching.

Fixes: #754