update openstack-ansible to current head of kilo for development
openstack-ansible hasn't been bumped forward since November. This
commit brings it forward to the head of osa kilo branch. Since we
only need a sha, and can't guarantee when osa will tag 11.2.7, it's
worth doing this now so we have a full 2 weeks to work with this
new sha and fix any issues that arise because of it, so that it can
be included in the r11.1.2 release.