Point master branch of rpc-openstack at head of liberty branch of OSA
In order to begin development towards a liberty release, we need to
switch our master branch to point at osa liberty. This commit bumps
forward to the head of osa liberty

12.0.5-13-ge002596 (12.0.5 plus 13 commits)

This commit also contains necessary changes to have the gate run and
pass linting, including:

- use ephemeral disk on cloud server when AIO
- bootstrap ansible before bootstrap aio (which in liberty is ansible)
- run lint using test-requirements.txt instead of dev-requirements.txt
- pass a list of locations that should be scanned for files containing
  information about what packages to build (allows us to define
  /opt/rpc-openstack/rpcd so that all of the rpc packages can be built at
  the same time by the osa repo-builder and we don't need separate rpc
  plays to build our own packages)
- remove unneeded rpc plays to build rpc repo
- correct the location of the horizon-manage.py file due to venv use
- pin setuptools==19.3 to allow holland to install
- pin elasticsearch<2.1.0 as this is a requirement of latest
  elasticsearch-curator
- slight var modification to drop the correct maas openrc file
- add swift as a pip package for maas - we need it on host for swift
  monitoring via swift-recon
- remove with_items usage from rpc_maas pip package installation to
  prevent it from expanding list items as variables (specifically swift)

Please note, this patch includes a change to the way rpc wheels are built.
Previously, we had plays in rpc-o to build wheels for rpc and place
them in a separate rpc repo. Because in liberty osa, yaprt has gone
away, we would have needed to rework those plays. There were 2 choices:

- rewrite the plays to use the new wheel building machinery
- utilise upstream patch https://review.openstack.org/#/c/271233/ to
  allow the osa wheel building process to build packages for us, and so
remove unneccessary duplication of plays in rpc-o.

I chose the latter, as it seemed the best choice to keep the work in
this patch minimal