Bump ceph role SHAs
This commit bumps SHAs for the ceph-common, ceph-mon, and ceph-osd
roles.

Currently, ceph-common will update ceph packages to the latest available
in the release's repo when the ceph playbooks are re-run.  While this
will not upgrade major ceph versions, this could be problematic and
should be avoided.

Bumping SHAs pulls in commit 9a6cfce in ceph-common, which changes the
default behaviour so that an explicit flag needs to be passed to ansible
if a deployer wishes to upgrade packages.