Refactor the disablement of aodh and ceilometer
The ability to override the value of DEPLOY_CEILOMETER is removed.
This variable is used by openstack-ansible scripts, currently
Ceilometer is not supported by rpc-o and so its value should always
be 'no'.

The removal of the ceilometer.yml and aodh.yml conf.d files is moved
to the AIO section, these files are added by the openstack-ansible
bootstrap-aio process which is not run for a customer deployment.
Removing these files means there are no hosts that are members of the
ceilometer aodh Ansible groups.

The env.d files are not removed, these files define the Ansible group
structure but have no effect without the corresponding conf.d files.

'rm' is changed to 'rm -f' so that the command does not error if the
file is missing.

There is no upgrade-specific requirement, beyond what deploy.sh does,
in terms of Ceilometer and Aodh and so the section has been removed
from the upgrade script.