Bump ceph SHAs
This bump is to include commits [1] and [2] which fix issue #879 which
was introduced by the last SHA bumps.  That issue is limited to OSD
hosts running 4 GB of RAM or below and was therefore not discovered in
our gating.

[1] https://github.com/ceph/ansible-ceph-common/commit/0bfa01a6cb5a934e4c2d203eeceb51af3c3de186
[2] https://github.com/ceph/ansible-ceph-common/commit/739230648581659a337902e7159c974547e94898

Closes issue #879

(cherry picked from commit 98ee984ad39769dfaacf1cc22674f5cf032774c9)