Enable Kilo to Liberty upgrades
Move openstack-ansible SHA to include the Liberty upgrade code.

Replace references to rpc_deploy directory with openstack_deploy.
rpc_deploy was a Juno-specific directory.

Fix the ansible-galaxy command in scripts/upgrade.sh by escaping the
newline character separating the two halves of the command.

Remove the directory /root/.pip from the hosts in the group
repo_container. The directory is created by the playbook
repo-pip-setup.yml on a Kilo deployment. It restricts the repo server to
using packages it already has access to which prevents the repo server
from upgrading based on Liberty requirements.