Use ansible_nodename for Cinder service checks
Replace ansible_hostname with ansible_nodename as the host argument
supplied to the cinder_service_check.py MaaS plugin by the configured
checks.

The default value for the hostname used in Cinder is supplied by
socket.gethostname. The Ansible fact that should be equivalent to this
is ansible_nodename.

ansible_hostname == ansible_nodename.split('.')[0]

Closes-issue: https://github.com/rcbops/rpc-openstack/issues/924