Prevent deployment of Ceilometer in AIO
Remove the Ceilometer conf.d file so that the Ceilometer groups are
empty thereby preventing the creation of any Ceilometer containers or
the Ceilometer playbooks finding any valid hosts.

Set swift_ceilometer_enabled to false so the role os_swift doesn't
attempt to integrate with Ceilometer.

This commit should have not impact on production deployments but is
intended to improve the the Kilo gating process.

Closes-issue: https://github.com/rcbops/rpc-openstack/issues/945