Constrain version of ansible-lint
ansible-lint version 2.5.0 causes scripts/linting-ansible.sh to fail.
The error appear to be due to a bug introduced into 2.5.0 that is
affecting tasks using the modules in rpcd/playbooks/library. There is an
issue tracking this on the ansible-lint GitHub repo [1].

This patch constrains the version of ansible-lint to avoid hitting the
bug. The version contraints mirror those in openstack-ansible Liberty.

[1] https://github.com/willthames/ansible-lint/issues/145

Closes-issue: https://github.com/rcbops/rpc-openstack/issues/963