Disable ceph-ansible fsid auto-generation
By default, ceph-ansible also creates a cluster fsid even when fsid is
overridden in user variables.  Commit [1] adds a new variable called
generate_fsid which, when set to false, will disable ceph-ansible's uuid
generation.

Note that leaving generate_fsid set to true does not have any effect on
the fsid we assign to the cluster, however it does create a file called
/etc/openstack_deploy/ceph_fetch/ceph_cluster_uuid.conf which does not
reflect the actual fsid of the cluster.

[1] https://github.com/ceph/ceph-ansible/commit/8e0d53e

Closes #1001