Deploy beaver only to hosts where pip is laid down
Beaver was trying to install to hosts where pip.conf was not being
laid down.  This caused beaver to fail to install because the host
didn't have the correct pip configuration to point at the rpc pip
repo, where the beaver packages are located.  This patch addresses
the problem by ensuring that we only install beaver on hosts that
have the correct pip.conf, and can therefore access the beaver
package from the repo.

Closes: #991