Add host security hardening functionality
openstack-ansible Liberty has backported the ability to use the role
openstack-ansible-security [1].

This commit updates the openstack-ansible submodule to include the new
security functionality. A new variable, DEPLOY_HARDENING, is added to
deploy.sh to determine whether this new functionality is applied to a
deployment. It is used to set the var apply_security_hardening in
user_variables.yml which controls whether or not the playbook
security-hardening.yml is run as part of setup-hosts.yml.

[1] https://git.openstack.org/openstack/openstack-ansible-security

Closes-issue: https://github.com/rcbops/rpc-openstack/issues/999