Change memcached monitoring to monitor curr_conns
Memcached previously monitored total connections, this is does not
tell us useful information, so I changed it to monitor current
connections and also alarm (WARNING) against reaching 90% of the
maximum active connection limit.

`memcached_connections` is also set globally so that it can be used
in both the memcached role and in the maas role.

Related: #1025