Install packages required by Paramiko 2.0.0
Paramiko recently released version 2.0.0, this is installed by
bootstrap-ansible.sh and causes the script to fail due to missing
dependencies. This is addressed in openstack-ansible kilo but is not
included in the latest tag. deploy.sh has been modified to install the
missing packages. This is required until the next openstack-ansible tag
is released.

Issue: https://github.com/rcbops/rpc-openstack/issues/1037