Alter pip behavior of horizon_extensions play
This fix alters the pip behavior of the `horizon_extensions` play to
bring it to liberty standards.  Since we are no longer running the
`repo-build` play from within the rpcd directory the existing
requirements file is no longer being used by the OSA `repo-build`
play.  This commit adds the `horizon_extensions_pip_packages` variable
to the defaults and uses the updated pip form for installation.

Closes-bug: #1100