Updated Keystone token overrides
This patch updates the keystone token overrides by changing keystone_token_provider
and keystone_token_driver in user_variables.yml to reflect the new entry point. This was done due to the fact
that the direct of import keystone.token.providers.uuid.Provider for the token provider
and keystone.token.persistence.backends.sql.Token for the token driver are deprecated
as of Liberty and may be removed in Newton.

Addresses #873