Fix horizon_extensions static folder permissions
When horizon-manage runs as root, the horizon-managed compressed files will end up
with root as owner in the static folder.

This commit fixes a consistency issue: All the static files served by the webserver should be owned by the webserver, and not root.

Issue: https://github.com/rcbops/rpc-openstack/issues/1158

Signed-off-by: Jean-Philippe Evrard jean-philippe.evrard@rackspace.co.uk