Use OSA code for building RPC horizon_extensions
This commit leverages OSA work instead of reinventing the wheel in RPC.

It uses OSA's repo-build process to create a wheel for horizon_extensions.
The wheel is then installed during RPC's horizon_extensions role.

No behavior was changed in the process: the wheel making still use master
branch of horizon-extensions repo, which should be moved to a sha in a later
commit for stable branches.

Issue: https://github.com/rcbops/rpc-openstack/issues/1100
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>